+++ main/models/wishlist_model.go
@@ -3,10 +3,10 @@ package models
 type WishlistModel struct {
 	Id        string
 	Name      string
-	Price     int64
-	Weight    uint
+	Price     uint64
+	Weight    uint32
 	Category  string
-	PerUnit   int
+	PerUnit   uint64
 	ImageUrl  interface{}
 	ProductId string
 	UserId    string
